requirements.txt
absl-py==1.4.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.5
dnspython==2.3.0
entrypoints==0.4
executing==0.8.3
filelock==3.8.0
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h11==0.13.0
h5py==3.8.0
httpie==3.2.1
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.10.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nest-asyncio==1.5.5
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
prompt-toolkit==3.0.29
protobuf==3.19.6
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
pyjokes==0.6.0
PyJWT==2.4.0
pymongo==4.3.3
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==2.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.1
pywin32==304
pyzmq==23.2.0
redis==4.3.4
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.5.1
rsa==4.9
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.8.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.2
stack-data==0.3.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.10.0
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.16.3
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
